Add a “PID” object to retrieve counter values for fixed processes
The native PDH API only supports the “Process” object that is identified by
instance#index, with and index that can change over time. This means that
it's not possible to get counter values for a given process without making
sure that the process has not a new index every time data has been collected.

This changeset abstracts this complexity by providing a “PID” object which is
backed by a “Process” object with two counters:
 - one for the actually desired value
 - one for the process id (“ID Process”).
Then, before the desired value is used, the process id is checked for change
and in case it's not the same, all processes are enumerated to find the new
instance#index for the process.